hw5-se-school_2022-architecture-design: refactored architecture with presentation and domain layers and created architecture design image
