+++ src/integration/java/com/example/btcsubscriber/IntegrationTest.java
@@ -1,6 +1,6 @@
 package com.example.btcsubscriber;
 
-import com.example.btcsubscriber.repository.SubscriptionsRepository;
+import com.example.btcsubscriber.subscription.SubscriptionsRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
