--- src/main/java/com/example/btcsubscriber/controller/ControllerExceptionHandler.java
@@ -1,16 +1,0 @@
-package com.example.btcsubscriber.controller;
-
-import com.example.btcsubscriber.exceptions.SubscriptionException;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class ControllerExceptionHandler {
-
-    @ExceptionHandler(SubscriptionException.class)
-    public ResponseEntity<String> handleCityNotFoundException(SubscriptionException exception) {
-        return ResponseEntity.status(409).body(exception.getMessage());
-    }
-
-}
