+++ src/main/java/com/example/btcsubscriber/email/EmailGenerator.java
@@ -1,0 +1,11 @@
+package com.example.btcsubscriber.email;
+
+import org.springframework.mail.SimpleMailMessage;
+
+import java.util.List;
+
+public interface EmailGenerator {
+
+  SimpleMailMessage generateEmail(long rate, List<String> recipients);
+
+}
