--- src/main/java/com/example/btcsubscriber/exceptions/RateException.java
@@ -1,7 +1,0 @@
-package com.example.btcsubscriber.exceptions;
-
-public class RateException extends Exception {
-    public RateException(String message) {
-        super(message);
-    }
-}
