--- src/main/java/com/example/btcsubscriber/exceptions/SubscriptionException.java
@@ -1,7 +1,0 @@
-package com.example.btcsubscriber.exceptions;
-
-public class SubscriptionException extends Exception {
-    public SubscriptionException(String message) {
-        super(message);
-    }
-}
