+++ src/main/java/com/example/btcsubscriber/rate/DefaultRateService.java
@@ -1,0 +1,23 @@
+package com.example.btcsubscriber.rate;
+
+import com.example.btcsubscriber.rate.provider.RateProvider;
+import lombok.RequiredArgsConstructor;
+
+import java.util.List;
+import java.util.Optional;
+
+@RequiredArgsConstructor
+public class DefaultRateService implements RateService {
+
+  private final List<RateProvider> rateProviders;
+
+  @Override
+  public Optional<Long> getRate() {
+    return rateProviders.stream()
+        .map(RateProvider::getRate)
+        .filter(Optional::isPresent)
+        .findFirst()
+        .orElseGet(Optional::empty);
+  }
+
+}
