+++ src/main/java/com/example/btcsubscriber/rate/RateException.java
@@ -1,0 +1,7 @@
+package com.example.btcsubscriber.rate;
+
+public class RateException extends Exception {
+    public RateException(String message) {
+        super(message);
+    }
+}
