+++ src/main/java/com/example/btcsubscriber/rate/RateService.java
@@ -1,0 +1,9 @@
+package com.example.btcsubscriber.rate;
+
+import java.util.Optional;
+
+public interface RateService {
+
+  Optional<Long> getRate();
+
+}
