+++ src/main/java/com/example/btcsubscriber/rate/provider/RateProvider.java
@@ -1,0 +1,11 @@
+package com.example.btcsubscriber.rate.provider;
+
+import java.util.Optional;
+
+public interface RateProvider {
+
+  Optional<Long> getRate();
+
+  ProviderType getType();
+
+}
