--- src/main/java/com/example/btcsubscriber/repository/SubscriptionsRepository.java
@@ -1,11 +1,0 @@
-package com.example.btcsubscriber.repository;
-
-import java.util.List;
-
-public interface SubscriptionsRepository {
-    void addSubscription(String email);
-
-    boolean isExist(String email);
-
-    List<String> getSubscriptions();
-}
