--- src/main/java/com/example/btcsubscriber/service/email/EmailGenerator.java
@@ -1,11 +1,0 @@
-package com.example.btcsubscriber.service.email;
-
-import org.springframework.mail.SimpleMailMessage;
-
-import java.util.List;
-
-public interface EmailGenerator {
-
-  SimpleMailMessage generateEmail(long rate, List<String> recipients);
-
-}
