--- src/main/java/com/example/btcsubscriber/service/rate/RateService.java
@@ -1,9 +1,0 @@
-package com.example.btcsubscriber.service.rate;
-
-import java.util.Optional;
-
-public interface RateService {
-
-  Optional<Long> getRate();
-
-}
