--- src/main/java/com/example/btcsubscriber/service/rate/provider/RateProvider.java
@@ -1,11 +1,0 @@
-package com.example.btcsubscriber.service.rate.provider;
-
-import java.util.Optional;
-
-public interface RateProvider {
-
-  Optional<Long> getRate();
-
-  ProviderType getType();
-
-}
