+++ src/main/java/com/example/btcsubscriber/subscription/SubscriptionException.java
@@ -1,0 +1,7 @@
+package com.example.btcsubscriber.subscription;
+
+public class SubscriptionException extends Exception {
+    public SubscriptionException(String message) {
+        super(message);
+    }
+}
